move some canvas error tests to new testfile
