+++ editor/src/components/canvas/ui-jsx-canvas-errors.spec.browser2.tsx
@@ -1,10 +1,82 @@
 import { BakedInStoryboardUID, BakedInStoryboardVariableName } from '../../core/model/scene-utils'
 import { act, render } from '@testing-library/react'
 import * as React from 'react'
+import { testCanvasErrorInline } from './ui-jsx-canvas.test-utils'
+import { TestAppUID, TestSceneUID } from './ui-jsx.test-utils'
 
-describe('UiJsxCanvas errors', () => {
-  it('handles a component that is not imported by throwing a ReferenceError', () => {
-    const renderResult = render(<div style={{ width: 50, height: 12 }}>{BakedInStoryboardUID}</div>)
-    expect(renderResult.baseElement.nodeName).toEqual('BODY')
+describe('UiJsxCanvas React errors', () => {
+  it('shows error when the React import is missing from storyboard', () => {
+    const canvasErrors = testCanvasErrorInline(
+      null,
+      `
+    import { View, Storyboard, Scene } from 'utopia-api'
+    import { Card } from '/card.js'
+
+    export var ${BakedInStoryboardVariableName} = (props) => {
+      return (
+        <Storyboard data-uid={'${BakedInStoryboardUID}'}>
+          <Scene
+            style={{ position: 'absolute', left: 0, top: 0, width: 400, height: 400 }}
+            data-uid={'${TestSceneUID}'}
+          >
+            <Card
+              data-uid='${TestAppUID}'
+              style={{ position: 'absolute', bottom: 0, left: 0, right: 0, top: 0 }}
+            />
+          </Scene>
+        </Storyboard>
+      )
+    }
+    `,
+      {
+        'card.js': `
+      import * as React from "react"
+      export var Card = (props) => {
+        return (
+          <div style={props.style}>This is a card</div>
+        )
+      }
+      `,
+      },
+    )
+    expect(canvasErrors.length).toEqual(1)
+    expect(canvasErrors[0].message).toEqual('React is not defined')
+  })
+  it('shows error when the React import is missing from card.js', () => {
+    const canvasErrors = testCanvasErrorInline(
+      null,
+      `
+    import * as React from "react"
+    import { View, Storyboard, Scene } from 'utopia-api'
+    import { Card } from '/card.js'
+
+    export var ${BakedInStoryboardVariableName} = (props) => {
+      return (
+        <Storyboard data-uid={'${BakedInStoryboardUID}'}>
+          <Scene
+            style={{ position: 'absolute', left: 0, top: 0, width: 400, height: 400 }}
+            data-uid={'${TestSceneUID}'}
+          >
+            <Card
+              data-uid='${TestAppUID}'
+              style={{ position: 'absolute', bottom: 0, left: 0, right: 0, top: 0 }}
+            />
+          </Scene>
+        </Storyboard>
+      )
+    }
+    `,
+      {
+        'card.js': `
+      export var Card = (props) => {
+        return (
+          <div style={props.style}>This is a card</div>
+        )
+      }
+      `,
+      },
+    )
+    expect(canvasErrors.length).toEqual(1)
+    expect(canvasErrors[0].message).toEqual('React is not defined')
   })
 })
