+++ editor/src/components/canvas/ui-jsx-canvas-errors.spec.tsx
@@ -696,98 +696,3 @@ export var ${BakedInStoryboardVariableName} = (props) => {
     `)
   })
 })
-
-describe('UiJsxCanvas React errors', () => {
-  it('shows error when the React import is missing from storyboard', () => {
-    const canvasErrors = testCanvasErrorInline(
-      null,
-      `
-    import { View, Storyboard, Scene } from 'utopia-api'
-    import { Card } from '/card.js'
-
-    export var ${BakedInStoryboardVariableName} = (props) => {
-      return (
-        <Storyboard data-uid={'${BakedInStoryboardUID}'}>
-          <Scene
-            style={{ position: 'absolute', left: 0, top: 0, width: 400, height: 400 }}
-            data-uid={'${TestSceneUID}'}
-          >
-            <Card
-              data-uid='${TestAppUID}'
-              style={{ position: 'absolute', bottom: 0, left: 0, right: 0, top: 0 }}
-            />
-          </Scene>
-        </Storyboard>
-      )
-    }
-    `,
-      {
-        'card.js': `
-      import * as React from "react"
-      export var Card = (props) => {
-        return (
-          <div style={props.style}>This is a card</div>
-        )
-      }
-      `,
-      },
-    )
-    expect(canvasErrors).toMatchInlineSnapshot(`
-      Array [
-        Object {
-          "columnNumber": undefined,
-          "lineNumber": undefined,
-          "message": "React is not defined",
-          "name": "ReferenceError",
-          "originalCode": undefined,
-        },
-      ]
-    `)
-  })
-  it('shows error when the React import is missing from card.js', () => {
-    const canvasErrors = testCanvasErrorInline(
-      null,
-      `
-    import * as React from "react"
-    import { View, Storyboard, Scene } from 'utopia-api'
-    import { Card } from '/card.js'
-
-    export var ${BakedInStoryboardVariableName} = (props) => {
-      return (
-        <Storyboard data-uid={'${BakedInStoryboardUID}'}>
-          <Scene
-            style={{ position: 'absolute', left: 0, top: 0, width: 400, height: 400 }}
-            data-uid={'${TestSceneUID}'}
-          >
-            <Card
-              data-uid='${TestAppUID}'
-              style={{ position: 'absolute', bottom: 0, left: 0, right: 0, top: 0 }}
-            />
-          </Scene>
-        </Storyboard>
-      )
-    }
-    `,
-      {
-        'card.js': `
-      export var Card = (props) => {
-        return (
-          <div style={props.style}>This is a card</div>
-        )
-      }
-      `,
-      },
-    )
-    expect(canvasErrors).toMatchInlineSnapshot(`
-      Array [
-        Object {
-          "columnNumber": undefined,
-          "lineNumber": undefined,
-          "message": "React is not defined",
-          "name": "ReferenceError",
-          "originalCode": undefined,
-        },
-      ]
-    `)
-  })
-})
