+++ editor/src/components/canvas/ui-jsx-canvas.test-utils.tsx
@@ -1,4 +1,4 @@
-import * as Prettier from 'prettier'
+import * as Prettier from 'prettier/standalone'
 import * as React from 'react'
 import { applyUIDMonkeyPatch } from '../../utils/canvas-react-utils'
 applyUIDMonkeyPatch()
